RAG_ESG/requirements.txt
langchain
faiss-cpu
Pillow
python-docx
pypdf
unstructured
torch
transformers
sentence-transformers
tqdm
tk
pyqt5
streamlit
pydantic
rank-bm25
jieba


